requirements.txt
flask
pandas==2.3.3
numpy==1.26.4
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
openpyxl==3.1.5
pillow
sentence-transformers
faiss-cpu
numexpr==2.10.2
bottleneck==1.5.0
transformers
google-genai
pydantic
cryptography
